+++ models/user.py
@@ -1,11 +1,13 @@
 from sqlalchemy import Column, INTEGER, VARCHAR, BOOLEAN
-from sqlalchemy.orm import declarative_base
+from sqlalchemy.orm import relationship
+from services.db.base import Base
 
 
-class User(declarative_base()):
+class User(Base):
     __tablename__ = 'User'
 
     id = Column(INTEGER, primary_key=True, autoincrement=True)
     username = Column(VARCHAR(255), nullable=False, unique=True)
     password = Column(VARCHAR(255), nullable=False)
     is_active = Column(BOOLEAN, nullable=False)
+    connections = relationship("Connection", secondary="User_Connection", back_populates="users")