+++ services/security/bearer.py
@@ -36,3 +36,4 @@ class Bearer:
                 status_code=status.HTTP_400_BAD_REQUEST,
                 detail="Invalid token"
             )
+